Armor renders model but dragon skull head thingy.
