+++ src/main/java/com/sekwah/narutomod/item/armor/NarutoArmorItem.java
@@ -1,8 +1,8 @@
 package com.sekwah.narutomod.item.armor;
 
 import com.sekwah.narutomod.NarutoMod;
+import com.sekwah.narutomod.client.renderer.NarutoRenderEvents;
 import com.sekwah.narutomod.item.interfaces.IShouldHideNameplate;
-import net.minecraft.client.Minecraft;
 import net.minecraft.client.model.HumanoidModel;
 import net.minecraft.client.renderer.BlockEntityWithoutLevelRenderer;
 import net.minecraft.world.entity.Entity;
@@ -54,18 +54,21 @@ public class NarutoArmorItem extends ArmorItem implements IShouldHideNameplate {
 
 
     // A little confusing but this is how it works now
-    /*@Override
+    @Override
     public void initializeClient(Consumer<IItemRenderProperties> consumer) {
         super.initializeClient(consumer);
         consumer.accept(new IItemRenderProperties() {
 
-            final BlockEntityWithoutLevelRenderer myRenderer
-                    = new NarutoItemRenderer(Minecraft.getInstance().getBlockEntityRenderDispatcher(), Minecraft.getInstance().getEntityModels());
+            @Override
+            public BlockEntityWithoutLevelRenderer getItemStackRenderer()
+            {
+                return NarutoRenderEvents.NARUTO_RENDERER;
+            }
 
             @Override
             public HumanoidModel getArmorModel(LivingEntity entityLiving, ItemStack itemStack, EquipmentSlot armorSlot, HumanoidModel _default) {
                 return armorModel;
             }
         });
-    }*/
+    }
 }
