--- src/main/java/com/sekwah/narutomod/item/armor/NarutoItemRenderer.java
@@ -1,12 +1,0 @@
-package com.sekwah.narutomod.item.armor;
-
-import net.minecraft.client.model.geom.EntityModelSet;
-import net.minecraft.client.renderer.BlockEntityWithoutLevelRenderer;
-import net.minecraft.client.renderer.blockentity.BlockEntityRenderDispatcher;
-
-public class NarutoItemRenderer extends BlockEntityWithoutLevelRenderer {
-
-    public NarutoItemRenderer(BlockEntityRenderDispatcher p_172550_, EntityModelSet p_172551_) {
-        super(p_172550_, p_172551_);
-    }
-}
