+++ src/main/java/com/sekwah/narutomod/util/ModelUtils.java
@@ -1,0 +1,26 @@
+package com.sekwah.narutomod.util;
+
+import net.minecraft.client.model.geom.PartPose;
+import net.minecraft.client.model.geom.builders.CubeListBuilder;
+import net.minecraft.client.model.geom.builders.MeshDefinition;
+import net.minecraft.client.model.geom.builders.PartDefinition;
+
+public class ModelUtils {
+
+    /**
+     * Solve some model tomfuckery and stop mc bitching
+      */
+    public static MeshDefinition createBlankHumanoidMesh() {
+        MeshDefinition meshdefinition = new MeshDefinition();
+        PartDefinition partdefinition = meshdefinition.getRoot();
+        partdefinition.addOrReplaceChild("head", CubeListBuilder.create(), PartPose.offset(0, 0, 0.0F));
+        partdefinition.addOrReplaceChild("hat", CubeListBuilder.create(), PartPose.offset(0, 0, 0.0F));
+        partdefinition.addOrReplaceChild("body", CubeListBuilder.create(), PartPose.offset(0, 0, 0.0F));
+        partdefinition.addOrReplaceChild("right_arm", CubeListBuilder.create(), PartPose.offset(0, 0, 0.0F));
+        partdefinition.addOrReplaceChild("left_arm", CubeListBuilder.create(), PartPose.offset(0, 0, 0.0F));
+        partdefinition.addOrReplaceChild("right_leg", CubeListBuilder.create(), PartPose.offset(0, 0, 0.0F));
+        partdefinition.addOrReplaceChild("left_leg", CubeListBuilder.create(), PartPose.offset(0, 0, 0.0F));
+        return meshdefinition;
+    }
+
+}
